% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in Fatality Analysis Reporting System data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{csv file containing data}
}
\value{
\code{fars_read} will search within the specified path for the 
  filename provided. If the file exists, it will be imported and returned as 
  a data frame table.  If it does not exist an error message will be returned, warning messages are suppressed
}
\description{
\code{fars_read} reads in the US National Highway Traffic Safety 
  Administration's Fatality Analysis Reporting System data.
}
\examples{
fars_read("data/fars_data.csv")

}
