% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Observations by Year (Old Dataset)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years relating to the file names to be read in}
}
\value{
\code{fars_summarize_years} will return a wide-formatted data frame.
}
\description{
\code{fars_summarize_years} will read in multiple Fatality Analysis Reporting 
  System data files based on the years provided and summarise the number of
  observations by month and year.
}
\examples{
fars_summarize_years(2015:2016)

}
\seealso{
\code{\link{fars_read_years}} to understand how the file name is created
}
