% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make File Name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{the year to add to the file name}
}
\value{
\code{make_file} will return a file name based on the year provided.
  For example, if 2023 is provided as the year the name that will be returned
  will be "accident_2023.csv.bz2".
}
\description{
\code{make_file} creates a name for the accident csv file based on the 
  year provided.
}
\examples{
make_filename(2023)

}
